对项目进行修改： 修改因为负pi和正pi重复，从而导致机动存在的问题 修改一些参数设置，将replay size和场景运行时长重新设计

+++ v1/src/red_swarm_policy/blue_rl/flight_quality.py
@@ -132,8 +132,9 @@ class FlightQualityTracker:
         maneuver_vectors = [np.array([command[0], command[1] * math.cos(command[2]) - 1.0,
                                       -command[1] * math.sin(command[2])]) for command in commands]
         opposite = sum(float(np.dot(a, b)) < 0 for a, b in zip(maneuver_vectors, maneuver_vectors[1:]))
-        left_right = sum(a[2] * b[2] < 0 for a, b in zip(commands, commands[1:]))
-        load_flips = sum(a[0] * b[0] < 0 or a[1] * b[1] < 0 for a, b in zip(commands, commands[1:]))
+        # Summing NumPy comparisons produces np.int64; reports need Python ints.
+        left_right = int(sum(a[2] * b[2] < 0 for a, b in zip(commands, commands[1:])))
+        load_flips = int(sum(a[0] * b[0] < 0 or a[1] * b[1] < 0 for a, b in zip(commands, commands[1:])))
         longest_extreme_run = current_extreme_run = 0
         previous_extreme: int | None = None
         for action, command in zip(actions, commands):
